refactor: 공통 모듈 상속 구조 도입으로 코드 중복 제거
